pyproject.toml
[project]
name = "mem-first-web-agent"
version = "0.1.0"
description = "Memory-first web agent -- Redis vector cache with Tavily web fallback"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.3.34",
    "langchain-core>=0.3.51",
    "openai>=1.68.2",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "markdownify>=0.14.1",
    "python-dotenv>=1.0.0",
    "numpy>=2.2.0",
    "tiktoken>=0.8.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
